Removed unexpected behavior of import lab method (#121)
When a lab with a specific title existed on a CML instance,
we joined that lab instead of creating a new lab with the same
title which would had been perfectly fine.  I believe that this
was some obsolete feature request from pre-CML 2.0 era where labs
possibly were enforced to have unique titles.

This feature requests was in my opinion by mistake delivered
in CML 2.7.0 and should be removed.  More importantly, it is
causing harm when working with deleted labs #120.